Code adaptation after tools update + inclusion of all tests in mypy + moved examples folder
