+++ examples/oscillator.py
