+++ tests/test_bouncing_ball_3d.py
@@ -15,7 +15,7 @@ from libcosimpy.CosimManipulator import CosimManipulator
 from libcosimpy.CosimObserver import CosimObserver
 from libcosimpy.CosimSlave import CosimLocalSlave
 
-from component_model.model import Model  # type: ignore
+from component_model.model import Model
 from component_model.utils.fmu import model_from_fmu
 
 
@@ -51,7 +51,7 @@ def bouncing_ball_fmu():
     build_path = Path.cwd()
     build_path.mkdir(exist_ok=True)
     fmu_path = Model.build(
-        str(Path(__file__).parent / "examples" / "bouncing_ball_3d.py"),
+        str(Path(__file__).parent.parent / "examples" / "bouncing_ball_3d.py"),
         project_files=[],
         dest=build_path,
     )
@@ -65,10 +65,7 @@ def test_bouncing_ball_class(show):
     If pytest is run from the command line, the current directory is the package root,
     but when it is run from the editor (__main__) it is run from /tests/.
     """
-    import sys
-
-    sys.path.insert(0, str(Path(__file__).parent / "examples"))
-    from bouncing_ball_3d import BouncingBall3D  # type: ignore
+    from examples.bouncing_ball_3d import BouncingBall3D
 
     bb = BouncingBall3D()
     result = []
@@ -432,13 +429,14 @@ def test_from_fmu(bouncing_ball_fmu):
 
 
 if __name__ == "__main__":
-    # retcode = pytest.main(["-rA", "-v", "--rootdir", "../", "--show", "False", __file__])
-    # assert retcode == 0, f"Non-zero return code {retcode}"
+    retcode = pytest.main(["-rA", "-v", "--rootdir", "../", "--show", "False", __file__])
+    assert retcode == 0, f"Non-zero return code {retcode}"
     # test_bouncing_ball_class(show=False)
-    Model.build(
-        str(Path(__file__).parent / "examples" / "bouncing_ball_3d.py"),
-        dest=(Path(__file__).parent / "test_working_directory"),
-    )
+    # Model.build(
+    #    str(Path(__file__).parent.parent / "examples" / "bouncing_ball_3d.py"),
+    #    dest=(Path(__file__).parent / "test_working_directory"),
+    # )
     # test_use_fmu( Path(__file__).parent / "test_working_directory" / "BouncingBall3D.fmu", False)
     # test_from_fmu( Path(__file__).parent / "test_working_directory" / "BouncingBall3D.fmu")
-    test_from_osp(Path(__file__).parent / "test_working_directory" / "BouncingBall3D.fmu")
+    # test_from_osp(Path(__file__).parent / "test_working_directory" / "BouncingBall3D.fmu")
+    # test_make_bouncing_ball(Path(__file__).parent / "test_working_directory" / "BouncingBall3D.fmu")
