+++ tests/test_model.py
@@ -26,7 +26,7 @@ def bouncing_ball_fmu(tmp_path_factory):
     build_path = Path.cwd()
     build_path.mkdir(exist_ok=True)
     fmu_path = Model.build(
-        Path(__file__).parent / "examples" / "bouncing_ball_3d.py",
+        Path(__file__).parent.parent / "examples" / "bouncing_ball_3d.py",
         project_files=[],
         dest=build_path,
     )
